Wrote the bulk of the code for XiCoin, but still a lot of work left to be done.
There's an issue with the way the difficulty is being raised, and it is currently impossible to go above 1 (I think this is due to how the SHA256 is being cast into binary for the difficulty check).

I'm purely checking in at this point so I have a backup in case I make a mistake.